+++ sphinx/source/changelog.rst
@@ -9,6 +9,8 @@ Changelog
 * Added support for :ref:`ssubBlockDirectives`
 * Added the ``--end`` option to :ref:`sna2skool.py` (for specifying the address
   at which to stop disassembling)
+* Added the ``--ctl-hex-lower`` option to :ref:`sna2skool.py` (for writing
+  addresses in lower case hexadecimal format in the generated control file)
 * Added the ``--hex-lower`` option to :ref:`skool2ctl.py` and
   :ref:`skool2sft.py` (for writing addresses in lower case hexadecimal format)